requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
plotly>=5.18.0
numpy>=1.24.0
transformers>=4.51.0
torch>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
